% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPalPicker.R
\name{funPalPicker}
\alias{funPalPicker}
\title{Interactively Select A Fun Palette}
\usage{
funPalPicker()
}
\description{
This function will open a Shiny gadget that will let you look at each
color palette.  You will see the palette and the hex values will be
printed out to the console on each selection.
}
\examples{
funPalPicker()
}
\author{
Seth Berry
}

