% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPals.R
\docType{data}
\name{funPals}
\alias{funPals}
\title{Hex Values for Palettes}
\format{A list containing 13 items
/describe{
 \item{paletteName}{The name of each palette}}}
\usage{
funPals
}
\description{
A list containing all of the palettes and their associated hex values.
}
\keyword{datasets}

