% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showFunPal.R
\name{showFunPal}
\alias{showFunPal}
\title{Select A Fun Palette}
\usage{
showFunPal(name, n)
}
\description{
This function will simultaneously plot the color palette and print
the hex values to console.
}
\examples{
showFunPal()
}
\author{
Seth Berry
}

